requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.33.3
libtorrent
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
pefile==2023.2.7
yara-python==4.3.1
requests==2.31.0
